# FIMEFavor.API(Controller)/FIMEFavor.API/bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FIMEFavor.API",
    "version": "v1"
  },
  "paths": {
    "/api/Categorias": {
      "get": {
        "tags": [
          "Categorias"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Categoria"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Categorias"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Categoria"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Categoria"
            }
          }
        }
      }
    },
    "/api/Categorias/{id}": {
      "get": {
        "tags": [
          "Categorias"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Categorias"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Categoria"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Categorias"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Categorias/{categoryId}/products": {
      "get": {
        "tags": [
          "Categorias"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "categoryId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ProductoYCategoriaBase"
              }
            }
          }
        }
      }
    },
    "/api/Cuentas": {
      "get": {
        "tags": [
          "Cuentas"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Cuentas"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Cuentas"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Cuentas"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Cuentas"
            }
          }
        }
      }
    },
    "/api/Cuentas/{id}": {
      "get": {
        "tags": [
          "Cuentas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Cuentas"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Cuentas"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Cuentas"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DetallesOrdenes": {
      "get": {
        "tags": [
          "DetallesOrdenes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/DetallesOrden"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "DetallesOrdenes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "agenteId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/DetallesOrden"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/DetallesOrdenes/{id}": {
      "get": {
        "tags": [
          "DetallesOrdenes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "put": {
        "tags": [
          "DetallesOrdenes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "DetallesOrdenes"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Orden/{id}": {
      "get": {
        "tags": [
          "Orden"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "ordenId",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Orden"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Orden"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Orden"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "Cuentas": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "matricula": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "correo": {
          "type": "string"
        },
        "contrasena": {
          "type": "string"
        }
      }
    },
    "Producto": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "modelo": {
          "type": "string"
        },
        "imagen": {
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "categoriaId": {
          "format": "int32",
          "type": "integer"
        },
        "categoria": {
          "$ref": "#/definitions/Categoria"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "existencia": {
          "format": "int32",
          "type": "integer"
        },
        "vendedorId": {
          "format": "int32",
          "type": "integer"
        },
        "cuentas": {
          "$ref": "#/definitions/Cuentas"
        }
      }
    },
    "Categoria": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "categoriaNombre": {
          "type": "string"
        },
        "productos": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Producto"
          }
        }
      }
    },
    "ProductoYCategoriaBase": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "productoId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "modelo": {
          "type": "string"
        },
        "imagen": {
          "type": "string"
        },
        "descripcion": {
          "type": "string"
        },
        "categoriaId": {
          "format": "int32",
          "type": "integer"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "existencia": {
          "format": "int32",
          "type": "integer"
        },
        "vendedorId": {
          "format": "int32",
          "type": "integer"
        },
        "categoriaNombre": {
          "type": "string"
        }
      }
    },
    "Orden": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "fecha": {
          "format": "date-time",
          "type": "string"
        },
        "clienteId": {
          "format": "int32",
          "type": "integer"
        },
        "cuentas": {
          "$ref": "#/definitions/Cuentas"
        },
        "total": {
          "format": "double",
          "type": "number"
        },
        "lugarDeEntrega": {
          "type": "string"
        },
        "repartidorId": {
          "format": "int32",
          "type": "integer"
        },
        "cuenta": {
          "$ref": "#/definitions/Cuentas"
        },
        "detallesOrdenes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/DetallesOrden"
          }
        }
      }
    },
    "DetallesOrden": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "productoId": {
          "format": "int32",
          "type": "integer"
        },
        "producto": {
          "$ref": "#/definitions/Producto"
        },
        "ordenId": {
          "format": "int32",
          "type": "integer"
        },
        "orden": {
          "$ref": "#/definitions/Orden"
        },
        "precio": {
          "format": "double",
          "type": "number"
        },
        "cantidad": {
          "format": "int32",
          "type": "integer"
        },
        "costoTotal": {
          "format": "double",
          "type": "number"
        }
      }
    },
...
  }
}